refactor(harness): extract artifact execution ports
